refactor : sortBySubWayRule() 메서드 리팩토링
